refactor: move DueDateCheckerTest to same package
(cherry picked from commit 354b63d2a8f6d3303a1df680091edbe2e7c751f5)

Conflicts:
  engine/src/test/java/io/camunda/zeebe/engine/processing/scheduled/DueDateCheckerTest.java

Resolved import conflicts only.
